Create geocoding service
